+++ src/controllers/commentController.ts
@@ -1,11 +1,16 @@
 import { RequestHandler } from "express";
-import { Types } from "mongoose";
 import { verifyToken } from "../middleware/authentication";
 import { getNotFoundErrorResponse } from "../middleware/errorHandler";
 import Comment from "../models/Comment";
+import { castObjectId } from "../utils/mongooseHelpers";
 
 const deleteCommentFromDatabase: RequestHandler = async (req, res, next) => {
-  const commentId = Types.ObjectId(req.params.commentid);
+  const commentId = castObjectId(req.params.commentid);
+
+  if (!commentId) {
+    return next();
+  }
+
   const comment = await Comment.findByIdAndDelete(commentId).catch(next);
 
   if (comment) {
